Refactor for tari_utilities moving some commonly used stuff into a single crate
