+++ base_layer/core/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.0.1"
 edition = "2018"
 
 [dependencies]
+tari_utilities = { path = "../../infrastructure/tari_util"}
 bitflags = "1.0.4"
 chrono = "0.4.6"
 derive = { path = "../../infrastructure/derive"}
