+++ base_layer/core/src/transaction.rs
@@ -30,17 +30,13 @@ use crate::{
 };
 
 use crate::types::SignatureHash;
-use crypto::{
-    challenge::Challenge,
-    commitment::HomomorphicCommitment,
-    common::{Blake256, ByteArray},
-    hash::Hashable,
-};
+use crypto::{challenge::Challenge, commitment::HomomorphicCommitment, common::Blake256};
 use curve25519_dalek::scalar::Scalar;
 use derive::HashableOrdering;
 use derive_error::Error;
 use digest::Digest;
 use std::cmp::Ordering;
+use tari_utilities::{byte_array::ByteArray, hash::Hashable};
 
 bitflags! {
     /// Options for a kernel's structure or use.
@@ -57,6 +53,8 @@ bitflags! {
         const COINBASE_OUTPUT = 0b00000001;
     }
 }
+
+type Hasher = Blake256;
 
 #[derive(Debug, PartialEq, Error)]
 pub enum TransactionError {
@@ -94,10 +92,8 @@ impl TransactionInput {
 
 /// Implement the canonical hashing function for TransactionInput for use in ordering
 impl Hashable for TransactionInput {
-    type Hasher = Blake256;
-
     fn hash(&self) -> Vec<u8> {
-        let mut hasher = Self::Hasher::new();
+        let mut hasher = Hasher::new();
         hasher.input(vec![self.features.bits]);
         hasher.input(self.commitment.to_bytes());
         hasher.result().to_vec()
@@ -137,10 +133,8 @@ impl TransactionOutput {
 
 /// Implement the canonical hashing function for TransactionOutput for use in ordering
 impl Hashable for TransactionOutput {
-    type Hasher = Blake256;
-
     fn hash(&self) -> Vec<u8> {
-        let mut hasher = Self::Hasher::new();
+        let mut hasher = Hasher::new();
         hasher.input(vec![self.features.bits]);
         hasher.input(self.commitment.to_bytes());
         hasher.input(self.proof.0);
@@ -215,10 +209,8 @@ impl TransactionKernel {
 
 /// Implement the canonical hashing function for TransactionKernel for use in ordering
 impl Hashable for TransactionKernel {
-    type Hasher = Blake256;
-
     fn hash(&self) -> Vec<u8> {
-        let mut hasher = Self::Hasher::new();
+        let mut hasher = Hasher::new();
         hasher.input(vec![self.features.bits]);
         hasher.input(self.fee.to_le_bytes());
         hasher.input(self.lock_height.to_le_bytes());
@@ -388,12 +380,13 @@ mod test {
     use crypto::{
         challenge::Challenge,
         commitment::HomomorphicCommitment,
-        common::{Blake256, ByteArray},
+        common::Blake256,
         keys::{PublicKey as PublicKeyTrait, SecretKey},
         ristretto::pedersen::DEFAULT_RISTRETTO_PEDERSON_BASE,
     };
     use curve25519_dalek::scalar::Scalar;
     use rand;
+    use tari_utilities::byte_array::ByteArray;
 
     #[test]
     fn build_transaction_test_and_validation() {
