+++ infrastructure/crypto/src/lib.rs
@@ -6,8 +6,6 @@ pub mod macros;
 pub mod challenge;
 pub mod commitment;
 pub mod common;
-pub mod hash;
-pub mod hex;
 pub mod keys;
 pub mod musig;
 pub mod signatures;
