+++ infrastructure/merklemountainrange/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.0.1"
 edition = "2018"
 
 [dependencies]
+tari_utilities = { path = "../tari_util"}
 derive-error = "0.0.4"
 digest = "0.8.0"
