+++ infrastructure/tari_util/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "tari_utilities"
+version = "0.0.1"
+edition = "2018"
+
+[dependencies]
+
+derive-error = "0.0.4"
+
+[dev-dependencies]
